mark google-cloud-cpp 2.2.1 as broken
The package changes ABI on each patch release (which are rare, but do
happen). The 2.2 release only pinned major and minor versions, this is
fixed for future versions, but the 2.2.1 release has the incorrect pinning
level.  Luckily there is nothing relevant for Conda in that release, all the
changes are related to Bazel-based builds.